refactor(pr-automation): split lifecycle validation responsibilities
